refactor: 컴포넌트 구조 변경
- 기존 container를 삭제하고 props로 처리
