+++ client/src/components/UI/organisms/index.ts
@@ -1,7 +1,5 @@
 export { ModalPopup } from './ModalPopup/ModalPopup';
 export { LectureList } from './LectureList/LectureList';
-export { BasketLectureListContainer } from './LectureList/BasketLectureListContainer';
-export { SearchedLectureListContainer } from './LectureList/SearchedLectureListContainer';
 export type { LectureListProps } from './LectureList/LectureList';
 export { Header } from './Header/Header';
 export { TimeTableMenu } from './TimeTableMenu/TimeTableMenu';
